Installed Angular
I had to Install Angular v16, because the I am not familiar with the new features of in angular v17. Angular v16 will help us to create mobile friendly application, and a great user experience across all platforms. Being that it is a well developed client side developer friendly technology